update dtype float32

+++ tests/test_wsireader.py
@@ -158,7 +158,7 @@ TEST_CASE_10 = [
     FILE_PATH,
     {"level": 2, "dtype": None},
     {"location": (0, 0), "size": (2, 1)},
-    np.array([[[239], [239]], [[239], [239]], [[239], [239]]], dtype=np.uint8),
+    np.array([[[239], [239]], [[239], [239]], [[239], [239]]], dtype=np.float32),
 ]
 
 # device tests
